Fire an init event when component is initialised
This event has been added so that the initialisation of this component
can be coordinated with another part of the JS system (the
chat-conversation module).

When the chat conversation module initialises we want it to fire an
event on the inset form module to set the appropriate state. However in
regular loading circumstances the chat conversation module will be
initialised before the form module has been initialised and thus the
event won't be listened to.

Firing this event allows the chat conversation module to wait until this
form is initialised before trying to make use of it.