Load a pending answer on page load
This configures the chat conversation module to be able to resume the
loading of a pending answer should the page be loaded with a pending
answer (for example if someone refreshes the page while waiting for an
answer).

The tricky thing about setting this up was dealing with the fact that
the form JS modules may not be initialised when this code executes. Thus we
need to either check that the form is already initialised (which depends
on window.GOVUK.Modules.start being executed) or set an event listener
on the form for it to fire a custom init event.