Add system tests for JS conversation features
This adds system tests for the key JS features used as part of the
conversation.

Unlike regular system tests it does utilise aspects that are not
explicitly on the happy path. I've done this because there is no other
form of integration testing for this functionality and the JS is
somewhat hard to test - so is a little contrived.

We may find that we don't think this level of testing is needed or
beneficial in time, particularly if these tests are slow or flaky.