Refactor prompt spec to use any amount of few shots
Previously this was built (probably by me) in an inflexible way that
was tied to there only ever being 3 few shots.

As the number of few shots will likely change at a reasonable frequency
and the few shots are structured in the YAML file in a consistent way,
it's pretty easy to use the few prompts in the file to structure the
test.

We can simply iterate over the few shots and map each assistant user
interaction into the desired format, then add them between
the system and rephrased question section of the prompt.