Move main element into conversation layout
This moves more HTML from the application layout and into the
conversation layout.

The motivation for moving main was that I wanted an additional element
that could be the container for a module that manages the onboarding
flow. I found I couldn't add a new element without causing layout
problems with flex so thought it was easy enough to just move the main
element and use that.