Ensure that QuestionsFilter behaves correctly with partial date params
This commit adds some logic to the validations in the QuestionsFilter form
to handle partial params being passed to the controller for start and end
date.

It also adds some request specs and updates the controller actions to
use the :unprocessible_entity status code when there are errors on the
form.