Add request tests for each component
This runs a basic test that checks that each component we've added can
produce a component-guide response that does not error.

The motivation for this has been discovering a few times that the
component guide is broken.

This doesn't catch issues such as the components looking wonky though
where we'd need a visual regression test, or accessibility warnings
which would require a JS enabled system test.