Move filterable table endpoint testing in shared example
There's tonnes of duplication between the admin questions (index) and
conversation show endpoint.

Both function almost identically with the exception of the conversation
index endpoint being scoped to a conversations question.

I've moved all the tests into a shared example and added a single extra
test to the conversations spec to test for the above.

One thing to note with the tests here is that the regexes that matches
the question messages are in the correct order:

```
/Hello world.*World hello/m
/World hello.*Hello world/m
```

require the /m flag to match correctly as the values are split over
multiple lines/rows in the table.