Update Tell me more button to post to limitations confirm
The tell me more button was previously a link that redirected to the
onboarding limitations page with a query param for more_information.

This updated the link to be a button with a name of more_information and
a value of true.

This now posts the to the limitations confirm endpoint. When
params[:more_information] is present, session[:more_information] is set
to true and the user is redirected to the limitations page where
they see the additional information.