Configure Opensearch indexes to use alias
It’s probably only a matter of time until we discover we have to do some
sort of OpenSearch reindex or index rebuilding. Using an alias as the
name will allow us to swap between any rebuilt indexes gracefully
without having to force application downtime.

Previously the index was named 'govuk_chat_chunked_content' in dev or
'govuk_chat_chunked_content_test' in test.

This commit updates the create_index method to use 'govuk_chunked_content_default'
on dev and 'govuk_chunked_content_default_test' on test as default and
allows you to specify a different name via an argument.

The index creates an alias of 'govuk_chat_chunked_content' in dev or
'govuk_chat_chunked_content_test' in test.

This alias continues to be used for all other calls to Opensearch.

One thing that is a bit awkward about this is that the get_mapping calls to Opensearch
using the indexes alias return an object that uses the indexes name as
the key.

For example say we name the new index 'govuk_chunked_content_default_v2',
the get mappings hash would return

```
{
  govuk_chunked_content_default_v2: {
    mappings: {
      properties: {
        ...
      }
    }
  }
}
```

We can get round this by updating the logic in the 'update_missing_mappings'
method to use the first value in the hash, then dig through the mappings.