Add close_openai_embedding spec helper
This commit adds a spec helper to solve the following problem:

How to create an embedding that is close to another embedding and will be found
by opensearch with a known score.

After a bit of experimentation - this code works.

It alters every element of the vector by the same amount (5%) - this causes it to
always have the same knn distance, regardless of the contents of the original vector.

This is required because, as the embeddings are random, we can never know what
the score between 2 embedding will be. This seems to work nicely.
In practice it means that the embedding it produces will always have a score
of 0.43589 when retrieved with the embedding given as a parameter.