Handle JSON requests for onboarding limitation endpoint
This commit updates the onboarding limitation endpoint to handle JSON
requests that will be made via our JS implementation.

This endpoint should only ever be hit via requesting a JSON response if
the user has already visited the page and clicked the 'Tell me more'
button.

This posts to the confirm_limitations endpoint which sets the
session[:more_information] to true.

At this point it redirects back to the limitations page.

Due to this if the session[:more_information] value has not been set
we should return a 'not_acceptable' status.

I've also added an error message for context, but i can remove that if
it's unecessary.

If the session[:more_information] value has been set we return JSON with
the following:

- a fragment which will be used to anchor the page to the additional
information
- a conversation_data hash with the onboarding module
- some html to append to the last message in the conversation. This is
the additional information that lives in the 'tell_me_more_messages'
partial
- Form html, that will replace th current form that uses the 'Tell me more'
button with a link to gov.uk