Add onboarding JS modules
Adds two new modules for onboarding.

The first one of these, ConversationOnboardingFlow, is a module that is
used to manage the transition between steps in the onboarding process.
It has the role of replacing HTML in the view, updating the URL,
updating the data attributes of the wrapper module.

The second one of these, Onboarding, is a module for handling an
onboarding view. It is to intercept a form submission and instead make
the submission via a fetch request. The JSON response of this response
is used to fire an event that ConversationOnboardingFlow can receive in
order to transition to the next view.

As we are not replacing all of the HTML that is used in the conversation
layout we need a mechanism to remove event listeners otherwise dead
modules will still listen to events. To do this I've set up an array on
the onboarding module that stores the event listeners so that these can
be removed when the module receives an event of `deinit`