Add CreateAnswerFeedback form and AnswerFeedback factory
This adds the form object that we use to create AnswerFeedback records.
It follows the same pattern established in the CreateQuestion form in that
it uses the #submit method to persist the record to the db provided it
is valid.

It also adds a the AnswerFeedback factory which can used to generate
records and adds a trait to the answer factory to generate an answer
with a feedback record.