Ensure conversation JS form doesn't prevent default for feedback
At the moment, when you submit the feedback form, the event listener
on the conversation form triggers a submit event listener. This
prevents the form being submitted and handles all the logic for
submitting a question.

We don't want to do this for feedback. This does the bare minumim to
ensure that the form is submitted when the one of the feedback buttons
is clicked and the feedback form is submitted.

It does not attempt to implement the JS implementation.