Add answer_feedback filter to the admin area
This updates the QuestionFilter to handle being able to filter
on whether the user indicated whether an answer was helpful or not.

I've defaulted the helpful boolean to true in the factory as it was
annoying not being able to use the :with_feedback trait in the answer
factory and being able to predict the distribution of the helpful
boolean. Mainly for the pagination tests.