Add onboarding JS specs
Adds specs for the both of the onboarding modules.

The `ConversationOnboardingFlow` module has been changed slightly to
refactor the scrolling code into its own function. It's very tricky to
test that an element has been scrolled into view (especially in this
case as the element is dynamically added to the DOM in the course of the
test).

So instead we'll assert that we call the `scrollIntoView` function of
the module, and verify that an element with the expect selector has been
passed in.