Stop retrying message processing when the index is not found.
In staging we have a problem where we are consuming messages from the publishing
API but something is wrong in the configuration and the index is missing.

With the current code this will be caught and retried indefinitely.
We don't have a delay on the retries and there's no point if the index is
missing, as they will never succeed.

We can revert this PR once we are sure all the indexes are in place - with the
correct names/aliases.