Scope submit event listener for questions to the form container
At the moment, the event listener that listens for form submission is
scoped to the module. This means that when the answer feedback forms
are submitted it triggers the event listener which was designed to
only trigger on question submission.

This commit moves the event listener to the form wrapper that contains
the question submission form.