Create canned answer with correct status when timed out
When the client polls the `GET answer` endpoint, we create a canned
answer if the question was created more than 2 minutes ago. This lets
the client inform the user that there's a timeout, so the user isn't
waiting for an answer forever and they can another question.

The canned response constants have been moved out of the
`OpenAIRagCompletion` class into the parent module as they're not really
to do with OpenAI.