Refactor the conversation-form component to have a wrapper
As we need an extra element outside of the component wrapper (the link)
we need it to be available as part of the component for JS behaviour (it
also feels like a smell to have a component having adjacent elements
outside the root). This PR adds a wrapper around the component so that
the component is all contained within one module.

This was a rather brutal exercise in shotgun surgery and has made me not
want to use a form as a root element of a component again. To
distinguish between the form and the wrapper I've used a terminology of
form, for the actual form element, and formComponent, for the wrapper
element.