Update survey link dynamically when question accepted
When we first render the form for a conversation a conversation does not
actually exist, it only exists when a user has asked their first
question. Once the conversation exists a user has a cookie of
conversation_id with the corresponding value.

This code adds a method call when the conversation-form receives a
question-accepted event to check whether we have this cookie and then to
update the survey link.