Remove superfluous specificity in system test
We don't need to be doing all of the within calls that we're making in
this suite, the only one that is actually required is for when checking
a question as we want to assert it exists in the conversation list and
not in the input field.

For those ones it was superfluous to specify :css as this is the default
selector.