Enforce unique index on question_id in answers table
We should ensure that it's not possible to create multiple answers for a
given question, which we can do by enforcing a unique index on the
`question_id` column on the `answers` table.

We rely on this unique constraint in `ComposeAnswerJob`. If a question
has timed out, we automatically create the answer record outside of this
job. If the job successfully finishes outside of this timeframe it'll
try to create the proper answer record, which we want to fail and raise
an `ActiveRecord::RecordNotUnique`. Adding this unique index ensures
that error will be raised.

There's no way to update an index to make it unique; it's got to be
removed and recreated.