Add JS and CSS for cookie banner
We've had the cookie banner component embedded in the HTML for this tool
for a while but not actually had a visible banner because we lacked the
JS and CSS. This adds it in.

Note: It is expected that the links in the banner don't work correctly,
they're not expected to work unless chat is on www.gov.uk.