Dismiss cookie banners for JS conversation system tests
The presence of the cookie banner causes the JS system tests involving
chat to fail in the CI environment (not managed to experience it locally
sadly). These additional steps dismiss the banner at the start of the
onboarding flow. Without this we see errors like:

```
  1) Conversation JavaScript features client side validation
     Failure/Error: click_on "I understand"

     Capybara::ElementNotFound:
       Unable to find visible link or button "I understand"

     [Screenshot Image]: /home/runner/work/govuk-chat/govuk-chat/tmp/capybara/failures_r_spec_example_groups_conversation_java_script_features_client_side_validation_320.png

     # ./spec/support/system_spec_helpers.rb:4:in `given_i_have_confirmed_i_understand_chat_risks'
     # ./spec/system/conversation_js_features_spec.rb:21:in `block (2 levels) in <top (required)>'
```

My theory is that the reason this is failing is because chromedriver
runs in a very small viewport and when it has a cookie banner the rather
delicate flex layout leads to the "I understand" buttons to no longer be
accessible.

I felt the simplest way to resolve this was to dismiss the banner,
rather than doing anything complex that involved clicking invisible
buttons. This has the advantage that any screenshots of the tests are
not overwhelmed by the banner.

I did worry that we might have to consider that the banner might not be
available, which would be rather painful as it'd be slow to wait for it
and then not have it available. However since headless chrome runs each
session without any cookie persistence we can expect it to always show
at the start of a spec run.

---

As an aside to see screenshots in CI I added this to the GitHub Action,
I wasn't sure that there would be ill consequences of having it
permanently so didn't add it in:

```
      - name: Upload screenshots
        uses: actions/upload-artifact@v4
        if: failure()
        with:
            name: capybara-snapshots
            path: tmp/capybara/
```