Update answer feedback from component to use fieldset
James has pointed out an issue with the current implementation in this
[PR discussion](https://github.com/alphagov/govuk-chat/pull/344#discussion_r1654939487).

Essentially screen readers can go into something called "form mode" and
ignore text that isn't directly related to the form. This means that
the text within the paragraph would be ignored.

To get round this we can use a fieldset. After playing around I've found
that floating the legend left displays inline with the body. Let me know
if there's a better way to do this.