Create shared views for privacy and limitations forms
In the controller actions for `privacy` and `limitations`, we render the
eponymous view for each action when requesting HTML, but we render the
same `components/onboarding_form` for a JSON request.

The HTML views wrap the onboarding views too; passing the form URL in.

The issue here is that we are configuring the URL in two places; one for
the HTML request and one for the JSON one. This means we need to keep
these in sync, and it's easy to change one but forget to change the
other.

So this commit intruduces a form view for each action where we can set
the URL in one place and render that view in both HTML and JSON
responses.