Refactor answer feedback component
The root of this component is now a div rather than a form - this is because when feedback is submitted, the form remained visible and the content inside it changed to "Thanks for your feedback. Hide this message". The hide button, however, isn't a submit button and no data is sent after clicking it i.e. we have a form with no submit button, which doesn't seem quite right semantically. Therefore, I changed this component to have a div as its root element so we can hide the form element after feedback submission and render the "Thank you" message outside of it.

Fix test