Update message-queue-consumption docs
I've attempted to document the process in went through in order to
populate production Opensearch index as we will need to follow the
same process when we want to reindex the data.

I've gone with not going into tonnes of detail about how to connect
to bash/rails console for govuk-chat and publishing-api since there's
plenty of documentation on that already.

I have added anything in that might be a gotcha (for example, needing
admin priviliges to view the RabbitMQ dashboard).