Add Opensearch healthcheck
We're adding some additional healthcheck#ready endpoint to ensure
that key infrastructure is working as expected. This commit adds one
for Opensearch.

We're checking the cluster health. If the cluster is green, we're returning
:ok, yellow :warning, and red :critical.

One thing that's worth noting is that we've had to add an attribute reader
to set a message that will be used to construct the JSON response.

The Checkup class in `govuk_app_config` looks to see is `message` has been
defined on the healthcheck instance. If defined then it is used to construct
the JSON response for that healthcheck

```
healthcheck_name: {
  status: instance.status,
  message: instance.message
}
```

If not then it only uses the status:

```
healthcheck_name: {
  status: instance.status,
}
```

This can be seen in this [bit of code (https://github.com/alphagov/govuk_app_config/blob/498f38bd25e276d755feae2d56eae4c12fcdb522/lib/govuk_app_config/govuk_healthcheck/checkup.rb#L54-L56).

By adding an attr_reader for the message and setting it when appropriate
we can get the additional information on why the endpoint is returning critical.

If the cluster health is red, we can see that is the issue and if any
other error is thrown, for example a connection error, we can see that
in the message.