Add OpenAI healthcheck
This adds a healthcheck to the "healthcheck/ready" endpoint that checks
our OpenAIClient is able to connect to the OpenAI API.

We've picked a simple endpoint to use for this healthcheck which is the
returns a list of available models.

If an error is thrown druing the status method we return :critical and
set the message attribute to the error message.