Fix undefined method edge case
This one is a bit of an edge case, but it was something that I
experienced when testing so it could happen in the real world.

It affects a non-JS request when asking a question that fails validation
immediately after asking a previous question.

The issue is that in the `update` controller action there's a
conditional checking if the question was valid. If it's not, and the
request is `html`, we render the `show` view. That view has another
conditional which, if true, calls this line:

```
@conversation_data_attributes[:"pending-answer-url"]
```

But the `@conversation_data_attributes` variable won't be set, so it
throws an error:

```
ActionView::Template::Error (undefined method `[]=' for nil)
```

This variable is set in the `show` action, which is why the exception
raising is never triggered there.