Add character-count component to conversation form
Adds the character-count component from the govuk-frontend component
system. It's configured to not show the user any feedback until they get
to ~50 characters from the max limit, which is 300 characters.

When they get within the threshold, they see "n characters remaining".
If they exceed the limit, they see "you have n characters too many".

If they submit the form after exceeding the limit, they don't see any
other feedback but the form submission stops so that we don't make a
request. Previously they'd see an error message above the input field in
this situation, but because the character-count component is already
showing them this feedback, we'll just leave that on screen.

A note on the `hint_id` change in `_conversation_form.html.erb`. The
element with this ID initally renders some (visually hidden) text of
"Please limit your question to #{maxlength} characters." When the
character-count module initialises, it needs a DOM element to target
where it'll render the "n characters remaining" text into. It makes
sense for us to use this DOM element. The component needs an element
with an ID of the input ID suffixed with `-info`[1] so that's why it was
changed here.

[1]
https://github.com/alphagov/govuk-frontend/blob/v5.4.0/packages/govuk-frontend/src/govuk/components/character-count/character-count.mjs#L132