Rework markup in conversation form
The markup was getting a little difficult to understand in this
component, so this refactors it a little.

The component renders the input field, button, error messages and
character count UI. The input field and button have a couple of
constraints:

1. They must be side-by-side
2. They must be the same height

This is relatively easy to do with flexbox, but there's a bit of
difficulty added by the character count component, which injects a DOM
element directly after the input field, over which we have no control.
This is compounded by the fact that the character count text is
displayed below the input field, whereas the error messages (no input,
etc) are shown above.

Previously the markup was set out in such a way that made it difficult
to visualise how the actual UI would look based on the code. This change
does the following:

1. Moves the error list to the top, outside of the flex container. This
   allows the errors to take up the full width of the outer container
2. Makes the input field wrapper and button wrapper the only two
   descendents of the middle flex container. This way it's easy to
   control their height and alignment relative to each other

Note: there's a little bit of undesirable nesting happening in the CSS
here with regards to the error list items. We have a nested rule like
so:

```css
.app-c-conversation-form__error-list {
  > li {
    margin-bottom: 0;
  }
}
```

Ideally we'd add a separate class to these `<li>` elements in the markup
and target that class in the CSS. The problem with doing that is that a
rule from the gov.uk component library takes precedence and applies a
5px bottom margin to the element. The rule in quesiton is:

```css
.govuk-list > li {
  margin-bottom: 5px;
}
```

In order to remove this bottom margin, which we need to do, we'd have to
do some awkward work to override this govuk-list style, which feels like
more effort than it's worth just to avoid a small bit of nesting.