Add JS loading states for questions and answers
This adds some loading feedback states when the user has JS enabled and
the question/answer responses are slow.

For the question - i.e. the request to the server when asking a question
- we show a "Loading your question..." message in the conversation list
if the request takes >500ms. Usually this request will be very quick so
we normally don't want to show the message as it's a bit jarring to
see the loading message get replaced immediately by the question the
user just asked.

For the answer, we show the loading message immediately. It almost
always takes a while to generate the answer as we're doing quite a bit
of work on the backend.

We do both loading states by including a `<template>` element in the
markup for both the question and answer, cloning the tempate and
appending it to the conversation list. When the loading is finished, we
just remove it from the DOM.