Evaluate the accuracy of the guardrails using exact match
Adapting the evaluation set from DS and adding which guardrails should fail, so we can
see exactly which guardrails are being hit/missed.

We want to use the entire output from the guardrail, not just the True/False
The numbers that follow indicate which guardrail has failed so we want to store
them. Also increasing max_tokens to 10 to allow for this.
We like to have some idea how the behaviour we want is performing, this gives us that.

Some of these examples look like the LLM got it - I labelled them based on the
guardrail name - they look like they should genuinely fail more than 1 guardrail.