Add Rake task for exporting evaluation data
The data science team want to be able to run auto-evaluation tests on
output from our application. To do that, they've got a predefined list
of test questions along with ideal answers and ideal sources.

So we need a way to export a bunch of data from our system so that they
can run it through their auto-evaluation pipeline.

This adds a new rake task (`evaluation:generate_report`) which runs
through a list of questions stored in the
`lib/data/evaluations/questions.yml` file and spits out a load of JSONL
with one line per question. The structure of the JSON documents in that
file has been agreed with the DS team.

The rake task can take a single optional argument, a path to a file to
output the JSONL to. If this is omitted, the JSONL is printed to
standard out.