Remove anchoring for conversations following the receipt of a new answer or feedback submission
Previously, users were anchored to the bottom of the viewport following the receipt of a new answer or submission of feedback. This means that they wouldn't see the flash notification/alert without scrolling all the way to the top. This commit removes anchoring so that the user can immediately see the flash notification/alert and a link is provided in the successful answer alert which will navigate users to their latest answer