Add AnswerComposition::OpenAIAnswer
This has been pulled from Kevin's gist here and tweaked based on feedback
https://github.com/alphagov/govuk-chat/pull/397/commits/96708860e5d24a3d058037a59708fad9b982e523

This is a base class that other OpenAI stragegies will inherit from. It's
key function is to provide a common interface for iterating through
the required steps to generate an answer.

When iterating through the steps, after each step is run if the context
object's aborted attribute is set to true, it breaks the loop and
returns the context's answer.