Fix broken character count
The character count was broken for two reasons:

1. GOV.UK Frontend changed the API in 5.0 to no longer use .init - easy
   fix
2. A bug in GOV.UK Frontend where global variables are overriden
   (https://github.com/alphagov/govuk-frontend/issues/5047)

To fix the global variable issue we need to require the file and then
store the global variable to a local variable before the global variable
is then overridden.

As a sprockets require only requires a file once, this problem could
easily reoccur if the file was required in another part of the Sprockets
build.