Add chunk columns to answer_sources table
This adds two new columns to the `answer_sources` table:
`content_chunk_id` and `content_chunk_digest`.

The ID field will allow us to store a reference to the chunk which
produced this answer source.

The digest field will tell us whether the chunk that produced the source
is still the same as the one stored in OpenSearch, i.e. it has or hasn't
changed since the `AnswerSource` record was created.

This commit sets the attributes when building the AnswerSource records.