Refactor building of AnswerSources
In the RAG completion class, we take all of the OpenSearch results and
build some `AnswerSource` records based on those.

Currently we group the results by `base_path` and only create one
source per base path. The `path` of the source is decided as so:

- If there is just one result for the base path, we use the source URL
- If there are multiple results, we use the base path

This commit moves the filtering to render time, rather than filtering
before saving to the database. So now we create one source per search
result, but what we show in the view is exactly the same as before
because we're still doing the filtering, just at a later point.

Keeping the 1-1 mapping of search results to answer sources means we can
build a more accurate picture of how the answer was generated for the
auto-evaluation. Currently we're actually losing some of the more
fine-grained data by creating fewer source records than we had search
results.

When we build the report for auto-evaluation, we can include all of the
original search results.