Add AnswerComposition::Pipeline::Context
This class uses the [interactor pattern](https://github.com/collectiveidea/interactor)
to provide a context object for a pipeline.

This object will be passed betweeen the steps each strategy requires to
generate an answer.

This class is responsible for holding the state of the pipeline and
will be reused across all OpenAI strategies.

it provides functionality to update the answer and assoications, store
various required attributes such as the question_message (message or
rephrased message).

It also provides a method to store the current status of the pipeline
so we know when to abort the processs and return the answer stored on the
context.

This commit also moves the functionality for building sources from the
OpenAIUnstructuredAnswer class to Answer model and adds some unit
tests.