Add base_path column to answer_sources table
We already store the full path against an AnswerSource, which we pull
from the content item in OpenSearch. We additionally want to store the
base_path as well, as we want to show that to the user in some
instances.

When an answer is generated using multiple sources, we group them by URL
so that we only present relevant, non-duplicated links to the user. We
do that by inspecting the base_path and the full path. If there's just
one result for the base path then we show the user the full path, as
that's more relevant. If there are >1 results which share the same
base_path then we show the user just that one base path.

So in order to implement this logic in the view, we need to store both
bits of data against the AnswerSource.