Add heading to answer_sources table
Currently we build up the title of the `AnswerSource` by taking the
content item's title and optionally appending a header to it if it's the
only content item for that given base_url for the source.

This concatenation is a view-layer responsibility though, rather than
something that should happen at the pre-database level.

To that end, we'll store the heading against the `AnswerSource` and use
that to build up the title we show in the view.