Update tests in openai_unstructured_answer_spec.rb
Apologies if this is a bit hard to parse, but it felt like the easiest way
to tackle these tests was to tackle them in one commit at the end as there
was no point in the process the majority of the tests were in a passing
state.

This was due to the compose answer functionality not being implemented.
The main changes made here are that i've stopped stubbing out the
QuestionRephraser since it handles updating the context and question_message.

There is some duplication in these tests since the steps are implicitly
being tested twice.