Add OpenAI stubs for structured answer generation
This adds some stubs that handle adding function calling functionlity
for structured answer generation.

It does this by adding a new stub_openai_chat_completion_structured_response
and extending the "stub_openai_chat_completion" stub to take an optional
"tool_calls" argument.

This is passed to the "response_body" method and is passed into the body.
I've chosen to extend the existing stubs rather than create a new one
because it doesn't require a massive amount of changes to existing stubs.

It might be that if we end up using lots of different function calls,
that we want to split these stubs out into separate methods. For now
i think this is probabably preferable though.