Refactor OpenAI Pipeline to pass pipeline to OpenAIAnswer
This commit refactors the OpenAI pipeline to pass the pipeline to the OpenAIAnswer
rather than the OpenAIUnstructuredAnswerComposer. It centres around 3 things:

1. updating the AnswerComposition::Composer to pass an array of pipeline
steps to the OpenAIAnswer class for the "open_ai_rag_completion" answer
strategy
2. updating the OpenAIAnswer class to accept the pipeline steps and iterate
through them in #call rather than in the #answer_pipeline private method
and adds unit tests
3. simplifies the OpenAIUnstructuredAnswerComposer by refactoring it into
a step which only:
 - makes the second OpenAI API call for the answer
 - generates an answer for a successful response

This implementation provides us with a few advantages:

- reduced inheritance between strategies
- more flexibility in the pipeline steps. We can now easily pass args into
each step. For example, we could pass the number of documents to retreive etc.
into the search results fetcher step. Allowing the same step to be used
in separate strategies with different configurations.
- easier to test the pipeline steps in isolation
- removes duplication in testing since the composer is now only
responsible for composing a successful answer