Configure Google Analytics
This uses the configuration information that Daniel provided for the ids
of production and integration.

It populates this data prior to calling the load-analytics script.

I don't really know how all our analytics works, however I'm expecting
that by configuring this we should get a page view event or similar that
can be used to confirm if analytics are working.