Add OpenAIStructuredAnswerComposer step
This commit adds a pipeline step for composing a structured input and
output for the second OpenAI API call we make to handle answer generation.

This step uses the structured system prompt and few-shots added in a previous
commit to construct the messages in the request. The main difference with
the messages are:

- we provide a schema for the contexts we're going to provide
- when we provide the contexts, we adhere to the schema

This request uses function calling to ensure that the llm responds with
structured output

```
tools: [
  {
    type: "function",
    function: {
      "name": "generate_answer_using_retrieved_contexts",
      "description": "Use the provided contexts to generate an answer to the question.",
      "parameters": {
        "type": output_schema[:type],
        "properties": output_schema[:properties],
      },
      "required": output_schema[:required],
    },
  },
],
tool_choice: "required",
```

We use the output schema Alessia has provided. The llm responds in JSON
which when parsed outputs in the following format:

```
{
  answer: "Lots of tax.",
  answered: true,
  call_for_action: "For more detailed information, you can visit [Pay more tax](https://www.gov.uk/pay-more-tax)",
  urls_used: ["/pay-more-tax#yes-really"],
}
````

This commit uses this output to construct the answers message. If the LLM
provides invalid JSON or the JSON is invalid against the schema we abort
the pipeline and return an answer with the status 'error_invalid_llm_response'
and a canned answer.

Further PRs will add functionality that:

- sets the status to 'abort_llm_cannot_answer' if it returns 'answered: false'
- add a new 'used' bool column to sources and mark unused sources as 'used: false'