Create a wrapper class for queue messages
This class is used to create a more intuitive API for working with the
queue messages. This is so that application code doesn't have to bear
the confusion that discarding a message actually means retry.

Working with the headers is a bit confusing because the
GovukMessageQueueConsumer::Message class has a method called headers to
represent Bunny::MessageProperties and should probably have been named
properties as headers is a concept of that class [1]. So to work with
headers you have to be accessing the headers property of headers.

[1]: https://github.com/ruby-amqp/bunny/blob/eb97f76e5fec33c1387baac4b83a54dae5a00f32/lib/bunny/message_properties.rb#L60-L62